Create a separate AnimationsManager for animations
