+++ src/app/services/three/scene-manager.ts
@@ -1,6 +1,5 @@
 import { Scene, Object3D, Color, LineSegments, Mesh, MeshPhongMaterial, LineBasicMaterial, Vector3, Group, AxesHelper, AmbientLight, DirectionalLight, Line, MeshBasicMaterial, Material, Points, PointsMaterial, MeshToonMaterial, Camera, BufferGeometry, TubeBufferGeometry, SphereGeometry } from 'three';
 import { Cut } from '../extras/cut.model';
-import * as TWEEN from '@tweenjs/tween.js';
 
 /**
  * Manager for managing functions of the three.js scene.
@@ -421,162 +420,6 @@ export class SceneManager {
         // Restoring to original position and then moving again with the current value.
         objectChild.position.divideScalar(previousScale).multiplyScalar(value);
       }
-    });
-  }
-
-  /**
-   * Animate the propagation and generation of event data.
-   * @param tweenDuration Duration of the animation tween.
-   * @param onEnd Function to call when all animations have ended.
-   */
-  public animateEvent(tweenDuration: number, onEnd?: () => void) {
-    const eventData = this.scene.getObjectByName(SceneManager.EVENT_DATA_ID);
-
-    const allTweens = [];
-    // Traverse over all event data
-    eventData.traverse((eventObject: any) => {
-      if (eventObject.geometry) {
-        // Animation for extrapolating tracks without changing scale
-        if (eventObject.name === 'Track') {
-          // Check if geometry drawRange count exists
-          let geometryPosCount = eventObject.geometry?.attributes?.position?.count;
-          if (geometryPosCount) {
-            // WORKAROUND
-            // Changing position count for TubeBufferGeometry because
-            // what we get is not the actual and it has Infinity drawRange count
-            if (eventObject.geometry instanceof TubeBufferGeometry) {
-              geometryPosCount *= 6;
-            }
-            if (eventObject.geometry instanceof BufferGeometry) {
-              const oldDrawRangeCount = eventObject.geometry.drawRange.count;
-              eventObject.geometry.setDrawRange(0, 0);
-              const eventObjectTween = new TWEEN.Tween(
-                eventObject.geometry.drawRange
-              ).to({
-                count: geometryPosCount
-              }, tweenDuration);
-              eventObjectTween.onComplete(() => {
-                eventObject.geometry.drawRange.count = oldDrawRangeCount;
-              });
-              allTweens.push(eventObjectTween);
-            }
-          }
-        }
-        // Animation for scaling out objects with or without position
-        else {
-          const hasPosition = !eventObject.position.equals(new Vector3(0, 0, 0));
-
-          const scaleTween = new TWEEN.Tween({
-            x: 0.01,
-            y: 0.01,
-            z: 0.01
-          }).to({
-            x: eventObject.scale.x,
-            y: eventObject.scale.y,
-            z: eventObject.scale.z
-          }, tweenDuration);
-          // Manually updating scale since we need to change position
-          scaleTween.onUpdate((updatedScale: Vector3) => {
-            const previousScale = eventObject.scale.x;
-            eventObject.scale.setScalar(updatedScale.x);
-            if (hasPosition) {
-              // Restoring to original position and then moving again with the current value
-              eventObject.position.divideScalar(previousScale)
-                .multiplyScalar(updatedScale.x);
-            }
-          });
-          allTweens.push(scaleTween);
-        }
-        // // Animation for moving objects outward with no position
-        // else if (eventObject.position.equals(new Vector3(0, 0, 0))) {
-        //     console.log(eventObject);
-        // }
-        // // Animation for moving objects outward
-        // else if (eventObject.position) {
-        //     const oldPosition = eventObject.position.clone();
-        //     eventObject.position.setScalar(0);
-        //     allTweens.push(new TWEEN.Tween(eventObject.position)
-        //         .to(oldPosition, tweenDuration));
-        // }
-      }
-    });
-
-    // Start all tweens
-    for (const tween of allTweens) {
-      tween.easing(TWEEN.Easing.Quartic.Out).start();
-    }
-
-    // Call onEnd when the last tween completes
-    allTweens[allTweens.length - 1].onComplete(onEnd);
-  }
-
-  /**
-   * Animate the collision of two particles.
-   * @param tweenDuration Duration of the particle collision animation tween.
-   * @param particleSize Size of the particles.
-   * @param distanceFromOrigin Distance of the particles (along z-axes) from origin.
-   * @param particleColor Color of the particles.
-   * @param onEnd Callback to call when the particle collision ends.
-   */
-  public collideParticles(
-    tweenDuration: number,
-    particleSize: number = 10,
-    distanceFromOrigin: number = 5000,
-    particleColor: Color = new Color(0xffffff),
-    onEnd?: () => void
-  ) {
-    const particleGeometry = new BufferGeometry().fromGeometry(
-      new SphereGeometry(particleSize, 32, 32)
-    );
-    const particleMaterial = new MeshBasicMaterial({
-      color: particleColor,
-      transparent: true,
-      opacity: 0
-    });
-    const particle1 = new Mesh(particleGeometry, particleMaterial);
-    const particle2 = particle1.clone();
-    particle1.position.setZ(distanceFromOrigin);
-    particle2.position.setZ(-distanceFromOrigin);
-
-    this.getScene().add(particle1, particle2);
-
-    const particles = [particle1, particle2];
-    const particleTweens = [];
-
-    const allEventData = this.getScene().getObjectByName(SceneManager.EVENT_DATA_ID);
-    allEventData.visible = false;
-
-    for (const particle of particles) {
-      new TWEEN.Tween(particle.material).to({
-        opacity: 1
-      }, 300).start();
-
-      const particleToOrigin = new TWEEN.Tween(particle.position).to({
-        z: 0
-      }, tweenDuration).start();
-
-      particleTweens.push(particleToOrigin);
-    }
-
-    particleTweens[0].onComplete(() => {
-      this.getScene().remove(particle1, particle2);
-      setTimeout(() => {
-        allEventData.visible = true;
-      }, 200);
-      onEnd?.();
-    });
-  }
-
-  /**
-   * Animate the propagation and generation of event data with particle collison.
-   * @param tweenDuration Duration of the animation tween.
-   * @param onEnd Function to call when all animations have ended.
-   */
-  public animateEventWithCollision(tweenDuration: number, onEnd?: () => void) {
-    const trackColor = this.getScene()
-      .getObjectByName('Track')['material']['color'];
-    this.collideParticles(1500, 10, 5000, trackColor, () => {
-      this.animateEvent(tweenDuration, onEnd);
     });
   }
 }
